Big update
I worked on the frontend for pokemon, the home-page, and the login page. Something weird happened so I had to adjust the height and width of the elden ring images. Please look into it when you can!

Also, I went ahead and added a back button to the header of pokemon and elden ring that takes you back to the home page. The login button now correctly takes you to the login page.

All buttons have a slight, slick animation to them now.

Please take a look at your respective parts and lets get everything looking cohesive for the final! Message me if you need anything - Will M